update strings dan ngasih konfirmasi logout dan konfirmasi hapus history

+++ app/src/main/java/pnj/pk/pareaipk/adapter/HistoryAdapter.kt
@@ -1,11 +1,13 @@
 package pnj.pk.pareaipk.adapter
 
+import android.app.AlertDialog
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import pnj.pk.pareaipk.R
 import pnj.pk.pareaipk.database.entity.HistoryEntity
 import pnj.pk.pareaipk.databinding.ItemScanBinding
 
@@ -31,9 +33,7 @@ class HistoryAdapter(
                 textTitle.text = scanHistory.class_label
                 textScanDate.text = scanHistory.scanDate
 
-                // Assuming you have a method to load image from URI
-                // image.loadImageFromUri(scanHistory.imageUri)
-
+                // Load image using Glide
                 scanHistory.imageUri?.let { uri ->
                     Glide.with(itemView.context)
                         .load(uri)
@@ -41,13 +41,26 @@ class HistoryAdapter(
                 }
 
                 buttonDelete.setOnClickListener {
-                    onDeleteClick(scanHistory)
+                    showDeleteConfirmationDialog(scanHistory)
                 }
 
                 buttonDetailHistory.setOnClickListener {
                     onDetailClick(scanHistory)
                 }
             }
+        }
+
+        private fun showDeleteConfirmationDialog(scanHistory: HistoryEntity) {
+            AlertDialog.Builder(itemView.context)
+                .setTitle(R.string.delete_confirmation_title)
+                .setMessage(R.string.delete_confirmation_message)
+                .setPositiveButton(R.string.delete_confirmation_positive) { _, _ ->
+                    // User confirmed, proceed with deletion
+                    onDeleteClick(scanHistory)
+                }
+                .setNegativeButton(R.string.delete_confirmation_negative, null) // Do nothing if canceled
+                .setIcon(R.drawable.ic_delete) // Optional: use your existing delete icon
+                .show()
         }
     }
